Remove timer from auth overlay
